fixed scroll issue for when plant gets added, editing styling
