+++ package.json
@@ -11,8 +11,10 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.8.1",
     "react-scripts": "5.0.1",
+    "react-scroll": "^1.8.9",
     "sass": "^1.58.0",
     "sass-loader": "^13.2.0",
+    "sweetalert2": "^11.7.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
