+++ src/style/partials/components/_buttons.scss
@@ -2,9 +2,9 @@
 // Buttons 
     .changePage{
         align-self: center;
-        width: 20%;
+        width: 100%;
         margin-bottom: 2rem;
-        padding: 8px 14px;
+        padding: 12px 30px;
         border: none;
         border-radius: 7px;
         color: $darkGray;
@@ -21,11 +21,12 @@
                 transition: 1s;
             }
     }
+
     .add{
         align-self: center;
         width: 30%;
         padding: 8px 14px;
-        margin-bottom: 2rem;
+        margin-bottom: 1.5rem;
         border: none;
         border-radius: 7px;
         color: $darkGray;
@@ -104,3 +105,29 @@
             }
     }
 
+
+    // Links
+        .backToHome{
+        display: flex;
+        justify-content: center;
+        align-self: center;
+        width: 30%;
+        padding: 8px 14px;
+        margin-top: 1.5rem;
+        border: none;
+        border-radius: 7px;
+        color: $darkGray;
+        background-color: $lightOlive;
+        font-size: 1.8rem;
+        font-weight: bold;
+        text-decoration: none;
+        box-shadow: -2px 2px 2px $lightTextShadow;
+        text-shadow: -0.6px 0.6px  $textShadow;
+        cursor: pointer;
+            &:hover, 
+            &:focus {
+                background-color: $lightGreen;
+                color: $white;
+                transition: 1s;
+            }
+    }
