+++ src/style/partials/pages/_contentPage.scss
@@ -11,7 +11,7 @@
       justify-content: center;
       width: 100vw;
       height: 55vh;
-      margin-bottom: 1vh;
+      margin-bottom: 1.5vh;
       text-shadow: -6px 5px 2px $textShadow;
       box-shadow: -1px 2px 4px $boxShadow;
       background-color: $darkGreen;
@@ -31,7 +31,7 @@
       justify-content: center;
       align-self: center;
       width: 100%;
-      height: 40vh;
+      height: 37vh;
       box-shadow: -4px 4px 6px $boxShadow;
       border-radius: 15px;
       border: 4px solid $darkGreen;
@@ -40,7 +40,7 @@
         justify-content: space-evenly;
       }
       input, select{
-        margin: 2rem 0 5rem 0;
+        margin: 2rem 0 3rem 0;
         border-radius: 10px;
       }
       label{
@@ -87,7 +87,8 @@
 
 footer {
   h4{
-    margin: 1rem 0;
+    margin-bottom: 1rem;
+    margin-top: 0;
     text-align: center;
   }
 }
