+++ src/style/partials/pages/_homePage.scss
@@ -11,8 +11,9 @@
     flex-direction: column;
     justify-content: center;
     text-align: center;
-    width: 100%;
+    width: 90%;
     height: 80vh;
+    border-radius: 200px 15px 100px 15px/15px 100px 15px 200px;
     color: $apricot;
     background-color: $lightOlive;
     text-shadow: -5px 4px 2px $lightTextShadow;
